refactor: CalculatorViewController 주석처리 및 위치 정리
